requirements.txt
Flask==3.0.0
openai>=1.0.0
stripe==7.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
